Reduce code to fewer files
